+++ MetadataExtractor.Tests/IO/SequentialAccessTestBase.cs
@@ -46,19 +46,19 @@ namespace MetadataExtractor.Tests.IO
         [Test]
         public virtual void TestGetInt8()
         {
-            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked((byte)0xFF) };
+            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked(0xFF) };
             var reader = CreateReader(buffer);
             Assert.AreEqual(unchecked(0), reader.GetInt8());
             Assert.AreEqual(unchecked(1), reader.GetInt8());
             Assert.AreEqual(unchecked(127), reader.GetInt8());
-            Assert.AreEqual(unchecked((byte)255), reader.GetInt8());
+            Assert.AreEqual(unchecked(255), reader.GetInt8());
         }
 
         /// <exception cref="System.IO.IOException"/>
         [Test]
         public virtual void TestGetUInt8()
         {
-            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked((byte)0xFF) };
+            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked(0xFF) };
             var reader = CreateReader(buffer);
             Assert.AreEqual(0, reader.GetUInt8());
             Assert.AreEqual(1, reader.GetUInt8());
@@ -87,7 +87,7 @@ namespace MetadataExtractor.Tests.IO
         public virtual void TestGetInt16()
         {
             Assert.AreEqual(-1, CreateReader(new[] { unchecked((byte)0xff), unchecked((byte)0xff) }).GetInt16());
-            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked((byte)0xFF) };
+            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked(0xFF) };
             var reader = CreateReader(buffer);
             Assert.AreEqual(0x0001, reader.GetInt16());
             Assert.AreEqual(0x7FFF, reader.GetInt16());
@@ -101,7 +101,7 @@ namespace MetadataExtractor.Tests.IO
         [Test]
         public virtual void TestGetUInt16()
         {
-            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked((byte)0xFF) };
+            var buffer = new byte[] { unchecked(0x00), unchecked(0x01), unchecked(0x7F), unchecked(0xFF) };
             var reader = CreateReader(buffer);
             Assert.AreEqual(unchecked(0x0001), reader.GetUInt16());
             Assert.AreEqual(unchecked(0x7FFF), reader.GetUInt16());
@@ -146,7 +146,7 @@ namespace MetadataExtractor.Tests.IO
         public virtual void TestGetUInt32()
         {
             Assert.AreEqual(4294967295L, (object)CreateReader(new[] { unchecked((byte)0xff), unchecked((byte)0xff), unchecked((byte)0xff), unchecked((byte)0xff) }).GetUInt32());
-            var buffer = new byte[] { unchecked((byte)0xFF), unchecked(0x00), unchecked(0x01), unchecked(0x02), unchecked(0x03), unchecked(0x04), unchecked(0x05), unchecked(0x06) };
+            var buffer = new byte[] { unchecked(0xFF), unchecked(0x00), unchecked(0x01), unchecked(0x02), unchecked(0x03), unchecked(0x04), unchecked(0x05), unchecked(0x06) };
             var reader = CreateReader(buffer);
             Assert.AreEqual(unchecked(0xFF000102L), (object)reader.GetUInt32());
             Assert.AreEqual(unchecked(0x03040506L), (object)reader.GetUInt32());
@@ -176,7 +176,7 @@ namespace MetadataExtractor.Tests.IO
         [Test]
         public virtual void TestGetInt64()
         {
-            var buffer = new byte[] { unchecked((byte)0xFF), unchecked(0x00), unchecked(0x01), unchecked(0x02), unchecked(0x03), unchecked(0x04), unchecked(0x05), unchecked(0x06), unchecked(0x07) };
+            var buffer = new byte[] { unchecked(0xFF), unchecked(0x00), unchecked(0x01), unchecked(0x02), unchecked(0x03), unchecked(0x04), unchecked(0x05), unchecked(0x06), unchecked(0x07) };
             var reader = CreateReader(buffer);
             Assert.AreEqual(unchecked((long)(0xFF00010203040506L)), (object)reader.GetInt64());
             reader = CreateReader(buffer);
@@ -205,7 +205,7 @@ namespace MetadataExtractor.Tests.IO
         {
             var nanBits = unchecked(0x7fc00000);
             Assert.IsTrue(float.IsNaN(BitConverter.ToSingle(BitConverter.GetBytes(nanBits), 0)));
-            var buffer = new byte[] { unchecked(0x7f), unchecked((byte)0xc0), unchecked(0x00), unchecked(0x00) };
+            var buffer = new byte[] { unchecked(0x7f), unchecked(0xc0), unchecked(0x00), unchecked(0x00) };
             var reader = CreateReader(buffer);
             Assert.IsTrue(float.IsNaN(reader.GetFloat32()));
         }
@@ -216,7 +216,7 @@ namespace MetadataExtractor.Tests.IO
         {
             var nanBits = unchecked((long)(0xfff0000000000001L));
             Assert.IsTrue(double.IsNaN(BitConverter.Int64BitsToDouble(nanBits)));
-            var buffer = new byte[] { unchecked((byte)0xff), unchecked((byte)0xf0), unchecked(0x00), unchecked(0x00), unchecked(0x00), unchecked(0x00), unchecked(0x00), unchecked(0x01) };
+            var buffer = new byte[] { unchecked(0xff), unchecked(0xf0), unchecked(0x00), unchecked(0x00), unchecked(0x00), unchecked(0x00), unchecked(0x00), unchecked(0x01) };
             var reader = CreateReader(buffer);
             Assert.IsTrue(double.IsNaN(reader.GetDouble64()));
         }
